Gather local settings into one file
