+++ src/actions/LoginActions.tsx
@@ -10,14 +10,7 @@ import { Airship, showError } from '../components/services/AirshipInstance'
 import { WalletCreateItem } from '../components/themed/WalletList'
 import { ENV } from '../env'
 import { lstrings } from '../locales/strings'
-import {
-  getLocalSettings,
-  getSyncedSettings,
-  LOCAL_ACCOUNT_DEFAULTS,
-  LOCAL_ACCOUNT_TYPES,
-  PASSWORD_RECOVERY_REMINDERS_SHOWN,
-  setLocalSettings
-} from '../modules/Core/Account/settings'
+import { getSyncedSettings, PASSWORD_RECOVERY_REMINDERS_SHOWN } from '../modules/Core/Account/settings'
 import { initialState as passwordReminderInitialState } from '../reducers/PasswordReminderReducer'
 import { AccountInitPayload } from '../reducers/scenes/SettingsReducer'
 import { config } from '../theme/appConfig'
@@ -30,6 +23,7 @@ import { runWithTimeout } from '../util/utils'
 import { loadAccountReferral, refreshAccountReferral } from './AccountReferralActions'
 import { getUniqueWalletName } from './CreateWalletActions'
 import { expiredFioNamesCheckDates } from './FioActions'
+import { LOCAL_ACCOUNT_DEFAULTS, LOCAL_ACCOUNT_TYPES, readLocalSettings, writeLocalSettings } from './LocalSettingsActions'
 import { registerNotificationsV2 } from './NotificationActions'
 import { trackAccountEvent } from './TrackingActions'
 import { updateWalletsRequest } from './WalletActions'
@@ -194,11 +188,11 @@ export function initializeAccount(navigation: NavigationBase, account: EdgeAccou
 
       accountInitObject = { ...accountInitObject, ...syncedSettings }
 
-      const loadedLocalSettings = await getLocalSettings(account)
+      const loadedLocalSettings = await readLocalSettings(account)
       const localSettings = { ...loadedLocalSettings }
       const mergedLocalSettings = mergeSettings(localSettings, LOCAL_ACCOUNT_DEFAULTS, LOCAL_ACCOUNT_TYPES)
       if (mergedLocalSettings.isOverwriteNeeded && syncedSettings != null) {
-        await setLocalSettings(account, syncedSettings)
+        await writeLocalSettings(account, syncedSettings)
       }
       accountInitObject = { ...accountInitObject, ...mergedLocalSettings.finalSettings }
