+++ src/actions/WalletListActions.tsx
@@ -7,29 +7,13 @@ import { ButtonsModal } from '../components/modals/ButtonsModal'
 import { SortOption } from '../components/modals/WalletListSortModal'
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { lstrings } from '../locales/strings'
-import { setAccountBalanceVisibility, setWalletsSort } from '../modules/Core/Account/settings'
+import { setWalletsSort } from '../modules/Core/Account/settings'
 import { GetState, ThunkAction } from '../types/reduxTypes'
 import { NavigationBase } from '../types/routerTypes'
 import { getCreateWalletType } from '../util/CurrencyInfoHelpers'
 import { parseDeepLink } from '../util/DeepLinkParser'
 import { logActivity } from '../util/logger'
 import { launchDeepLink } from './DeepLinkingActions'
-
-export function toggleAccountBalanceVisibility(): ThunkAction<void> {
-  return (dispatch, getState) => {
-    const state = getState()
-    const { account } = state.core
-    const currentAccountBalanceVisibility = state.ui.settings.isAccountBalanceVisible
-    setAccountBalanceVisibility(account, !currentAccountBalanceVisibility)
-      .then(() => {
-        dispatch({
-          type: 'UI/SETTINGS/SET_ACCOUNT_BALANCE_VISIBILITY',
-          data: { isAccountBalanceVisible: !currentAccountBalanceVisibility }
-        })
-      })
-      .catch(showError)
-  }
-}
 
 export function updateWalletsSort(walletsSort: SortOption): ThunkAction<void> {
   return (dispatch, getState) => {
