+++ src/components/services/PasswordReminderService.ts
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { setPasswordReminder } from '../../actions/PasswordReminderActions'
+import { setPasswordReminder } from '../../actions/LocalSettingsActions'
 import { connect } from '../../types/reactRedux'
 import { PasswordReminder } from '../../types/types'
 import { matchJson } from '../../util/matchJson'
