+++ src/components/themed/TransactionListTop.tsx
@@ -6,7 +6,7 @@ import AntDesignIcon from 'react-native-vector-icons/AntDesign'
 import Ionicons from 'react-native-vector-icons/Ionicons'
 import { sprintf } from 'sprintf-js'
 
-import { toggleAccountBalanceVisibility } from '../../actions/WalletListActions'
+import { toggleAccountBalanceVisibility } from '../../actions/LocalSettingsActions'
 import { Fontello } from '../../assets/vector'
 import { getSymbolFromCurrency, SPECIAL_CURRENCY_INFO, STAKING_BALANCES } from '../../constants/WalletAndCurrencyConstants'
 import { ENV } from '../../env'
